Update WaitForInstancesSignal, remove WaitForInstancesStopped (#29)
WaitForInstancesSignal now can wait for an instance to stop or for for specific serial output data.
CreateInstances automatically streams serial port 1 to GCS

If SerialOutput is set the step will not complete until a line in the serial output matches
SuccessMatch or FailureMatch. A match with FailureMatch will cause the step
to fail.