translate_el: use python-guestfs
This tool should mount the disk independently if LVM is being used or if
/boot is in a separate partition and perform customizations in the VM
image.

Also, I tested with guestmount tool, but it seems that running
grub2-mkconfig through chroot makes the roofs to point to /dev/fuse
instead of the disk, so a full virtualisation should be used. One of the
options is to use virt-customize, but I find using python-guest to be
better, since it is not bash, we can handle errors better and have more
interoperability.