Clean up for Anaconda log saving on *EL7. (#228)
On EL7-based distros, Anaconda writes /var/log/anaconda and /root/*-ks.cfg
*after* all of the postinstall kickstart scripts have run. This means it's
not possible to remove all of these logs in a postinstall script. Instead,
the inst.nosave=all kernel param triggers it to avoid saving logs.

This change enables this parameter by default and adds an el_savelogs
option to the workflows to disable this for debugging purposes.

The relevant postinstall scripts are also fixed to remove the now unnecessary
cleanup.