Update tables to make them render properly
When I view this page, https://googlecloudplatform.github.io/compute-image-tools/image-import.html#compatibility-and-known-limitations , the tables for "Windows" and "Linux" in the "Compatibility and Known Limitations" section are not appearing as tables. They appear as a long string of text, which is difficult to read.

Perhaps it is because "There must be at least 3 dashes separating each header cell" (from https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet#tables).