image_test: stop sshguard
If there is a sshguard service, disable it.
When we remove enable-oslogin from metadata or set it to false, the
tests will try to perform an ssh connection to verify if the access was
revoked accordingly.
But attempting to do an ssh connection when expected to fail causes the
sshguard to increase the level of dangerousness of a given IP, and
eventually sshguard blocks this IP it it reaches a given threashhold.