image_test: replace FULLNAME by RealName
The RealName of a resource is composed by the name of the parent
workflows, but the genName function truncates it if too big, and it
can do a different truncation per resource, which means that using
FULLNAME variable is not reliable to deduce the RealName of the
resources.

It becomes a problem when we have several subworkflows where the
RealName of the resources are bigger, which breaks usages like the
oslogin test, where we need to pass the RealName of an instance
to another using metadata.