Fix StopInstances serial output log
When a machine was on "STOPPING" status, the SerialOutput would
eventually throw the warning `googleapi: Error 503: TIMEOUT,
backendError`, which is not really helpful as it's expected that the
SerialOutput will be closed soon.

By considering the "STOPPING" instance status an expected reason for
having the serial output closed, this warning is suppressed.