Avoid rebooting on boot test without catching output
Fast instances could manage to print the boot message and restart
without daisy noticing it. When that happend, only the "REBOOT" was
caught but before a "BOOT", which led to a test filure by timeout.

This patch used daisy to control the rebooting which should make this
behavior not happening anymore.