metadata-scripts test: improve post-stop string matching
There is a StopInstances and a WaitForInstanceSignal running in
parallel. The serial is not possible to be read very quickly after the
StopInstances step starts so having 3 WaitForInstanceSignal depending on
each other led to at least one of them running after the serial was
closed, which made the test to fail.

However having 3 WaitForInstanceSignal polling every 0.5s will blow your
API request per seconds quota so I reduced a little bit by increasing
the interval for the opposite test that is checking the serial output on
start up (no hurries on that one).