oslogin-ssh test: Avoid empty ssh-keys inclusion
The old method was stacking a '\n' at the ssh-keys metadata list when it
was removing the last key (it could clear a '\n' at the end at removal,
but only after removing, the new list had a '\n' at the end).

This fix will avoid the project metadata being overwhelmed by empty keys
after several runs and it will even clear them, if it happens again.