Reduce boot, disk and multi-nic test conflicts (#526)
Boot test could fail as the "BOOTED" message was overwritten by other
boot messages on rsyslog (race condition). Adding a 10s sleep should
avoid this kind of issue

Disk and multi-nic tests uses a lot of a project quota for testing
several images. As the bottle necks for completing the tests are the
oslogin tests, running disk/multi-nic tests one at a time will not
impact total time for the tests run, but will require a lower resource
quota.