Fix configuration test for centos/rhel
When verifying if a package is installed or not, the command
'yum --assumeno install package_name' returns 1 insetad of 0 if there is an
update to the package available, even if the package is installed, causing the
test to fail.

This patch replaces this commad by 'yum list installed package_name'.