Perform regex matching instead of plain string matching
In some cases, assertion on an output in serial console becomes tricky
with plain strings.
For example:
to check if a package is installed with rpmquery on centos, i am using
rpmquery -a {packagename}
The result is {packageName}-{version}-{arch}.
Assertion on packagename is ok but this might introduce a bug where
packge is not installed and output is
"package {packagename} is not installed".