Test new export flow: e2e tests to verify 10G/210G/1T/8T (#713)
There are overall 3 changes:

1. Improvement of workflow: delete disk resource proactively. Without this change, disk resource will keep alive during the whole testing, which may exhaust quota.
2. Test preparation workflow: it's done by files started with "create_random_disk_*". It creates a image in project "compute-image-tools-test" with full-filled random data, which will be used in the e2e test. It's a one-time preparation. We don't want to populate random data during the e2e test runnung, because it's quite time consuming.
3. Actual e2e test workflow: it exports the image to vmdk format, and then verifies by import && mount && ls.
This test set covers size of 10GB (fast and easy), 210GB (doesn't work in old workflow but supposely work fine in new workflow), 1TB (a reasonable large disk), and 8TB (a large disk that we theoretically can support).

Exporting a large image without format converting is not covered in this CL due that logic is not changed recently. I will cover it later as a lower priority.