answer yes to all questions (#750)
when a new yum repository is added by the user, we add the repo
details in the yum database, but do not import the gpg keys.
When we check for updates, yum will first try to add the gpg keys
for the newly added repo. Yum prompts for confirmation to add
the gpg keys.
Without this change, yum update check fails.