[export cli tool]Exporter refactoring (#860)
This is a preparation for import/export API. It's a refactoring without any feature changes.

It moves logic of export (and unit tests) to "exporter.go" from "main.go". The purpose is to make exporter a library so API can call.
It renamed a couple of package name under "cli_tools". The reason is that google3/third_party is incompatible with package names different from its folder name.