A couple of improvements for gce_export (#863)
gce_export responses for exporting raw format. A couple of improvements aree done by this PR.

Reliability: chunk size
Removed chunkSize=0, which means, using default setting instead. By testing, without this change, while exporting by a non-gcp computer, for a 2GB OS disk, it's 33% chance to fail; for a full-filled 10GB disk, it's 100% to fail. It's very fragile because it's very hard to finish sending a huge single request (256MB) via public network. Theoretically, even for a gcp computer, it is not super stable as well. So I changed it to default chunk size. With this change, it's 100% success. The perf is almost not impacted: 2s slower for a 10GB disk, which is, 1m44s v.s. 1m42s.

Reliability: retry logic
Moved "defer os.Remove(in)" to the correct place. Without this change retry logic doesn't work: once any error happens, os.Remove(in) will be executed in the for-loop. In that case, the next retry will fail because the file has been removed. I moved it to the "break" place and verified it works.

Bug: Fixed how to construct the file path for "_part" file. Before the fix, the expected path "/upload/abcde_part1" will become to "/uploadabcde_part1".

Concurrent issue: for the composing, I added a random prefix for "compose_i" file. It's because when user run multiple exports concurrently targeting the same file name, the dup operation happens to "compose_i" file and can fail the whole workflow. It can be repo by my test. The scenario can happen in the real world: customer may run dup commands by accident, or run a new command but forgot to kill the previous command.

Bug: Missed use of parameter "oauth": Usage of this parameter was deleted in the last version. I added it back.