Fix OVF import failure when location restrictions are on (#895)
Fix OVF import failure when location restrictions are on due to images being created in multi-regions. Since this change is internal to the importer, added releaseTrack flag to indicate which Compute API release track is allowed to be used. Image location is only in Beta and we shouldn't use Beta API in GA release.