daisy: compute: Retry on transport-level failures (#919)
* daisy: compute: Retry on transport-level failures

Currently, Daisy only retries when a GCE API error returns a
googleapi.Error. However, errors that are not googleapi.Errors are most
likely transport errors, which should always be retried. Lets retry for
errors that are not googleapi.Errors as well.

* Only retry on specific transport errors.

Instead of retrying on all unknown errors, only retry on "connection
reset by peer" and "unexpected EOF" errors.