Log wrapper usage/debug info to logging service (#915)
Log wrapper usage/debug info to Firelog.
There may be more change after the privacy review and the onboarding process of sawmill/firelog.

It defines structs of Firelog request (which will be converted to json and sent).
It adds logic in image/instance import/export main function to fill the log params.
It handles how to honor Firelog response and handle it correctly.
It fixes some minor things, for example: remove "Privacy" tag from log.

Core logic of the PR is in "firelog_logger.go"
Now the log logic is disabled by "serverLogEnabled = false" in firelog_logger.go, so it won't be sent to server. We will enable it when server side is ready.