If translation fails, alert user of what was detected on disk. (#1206)
This adds a new module, cli_tools/common/distro, that is used to determine whether the detection result matches the user's input. This will get used in the upcoming OS detection project.

Notably, the module doesn't (yet) support Windows. That's a tactical choice, since currently we only perform detection on Linux, and unlike Linux, in Windows the user-facing version strings are different than the internal version strings [1]. So, I'm leaving implementation of Windows for when it's needed, as that will help define the interface.

Lastly, if the detection result doesn't match the user's input, then we replace the current error with a new one. I like this approach, since it improves the signal-to-noise ratio, and since the Daisy error is logged a few lines before.