SUSE import: Don't install cloud-init; require iputils for tests. (#1214)
This was motivated by failed SUSE 15.1 integ tests. The root cause was that cloud-init was marked as optional, and when it wasn't installed, the iputils transitive dependency wasn't added. The integration test expected that ping was available, so when cloud-init isn't installed, ping wasn't available (assuming the original image didn't have it).