Validate arguments prior to running import. (#1217)
(Refactoring leading toward splitting inflate and import workflows)

This further pulls dependencies out of the importer package by introducing an args package that is fully responsible for parsing, validating, and populating arguments. This introduces an invariant that once importer.Run is invoked, the arguments are known to be valid.

This allows us to have different importers.