daisy/: added variable parsing to resize-disk. (#1280)
Currently, the resize-disk commmand can only take integer input and
cannot parse variables in a workflow. In order to allow variables in
this command, a new field "SizeGbVar string" is added, and then its
value is converted into int64 and stored in the original "SizeGb int64"
field. In this way, everything else except the ResizeDisk struct fields and
populate function remains untouched.