An image import suite focused on translation logic. (#1277)
Historically, the majority of import tests are run as daisy integration tests, and therefore bypass the wrapper logic. This change adds a suite to the cli_tools_e2e_test framework that focuses on import logic. The existing e2e suite is recharacterized as "CLI tests", since it focuses on the API contracts of the CLI tools.

Key features, compared to existing solutions for image import tests:

* Support for negative tests, which verify the output of error messages
* Efficient test representation (compared to an entire JSON document)