Avoid resizing PDs on inflation worker during import. (#1283)
We have observed import failures resulting from a lack of permissions on the inflater worker. Specifically a lack of permission to resize PDs.

This change pre-inspects the virtual disk file to find its physical and virtual size, and uses those values when creating the scratch and inflation disk, respectively.

It includes a fallback mode for inspection failures: if inspection fails, then the pre-existing default values of 10GB are used.