Load inspection workflow relative to workflow root directory. (#1294)
The `gcr.io/compute-image-tools/gce_vm_image_import:latest` docker image was at runtime with an error of:

```
[import-image]: 2020-07-20T19:15:30.92Z open daisy_workflows/image_import/inspection/inspect-disk.wf.json: no such file or directory
```

This occurred since loading of `daisy_workflows/image_import/inspection/inspect-disk.wf.json` assumed that the directory `daisy_workflows` was available in the process's working directory. This isn't true for the  `gcr.io/compute-image-tools/gce_vm_image_import:latest`  image. It has a directory layout of:

```
/gce_vm_image_import
/daisy_workflows
/workspace
...
```

When Cloud Build invokes `gce_vm_image_import`, it sets the working directory to  `/workspace`, and the daisy_workflows directory is not a child of the current working directory. [1]

1. https://cloud.google.com/cloud-build/docs/build-config