Ensure that gcloud shows image import translation logs. (#1299)
Since #1227, we've been using modularized daisy workflows in image import. A side effect is that gcloud isn't showing all of the log lines to users. This was caused by:

1. gcloud is configured to show lines that start with [image (or one of a few other prefixes)
2. daisy emits log lines using the workflow's name as a prefix
3. Some of the modularized workflow names don't start with image.

This PR is a quick fix that dynamically renames the workflows to have a name that matches the prefix. I have b/161567644 to track a longer-term fix where gcloud output isn't coupled to workflow names.