Update translate.ps1
Restricting to a single reboot if the metadata variables are not available. If the variables are still unavailable after the metadata reboot the import will fail with the specific error.

In further investigating another common is the imports had VMWare tools removed and they were rebooted. So like the other scripts added 1 minute delay as the delay may likely solve the problem. Also since this introduces a minute delay, reducing the sleep between attemps to 1 second.