Update translate.ps1
Chatted with Eric and he made a great point that the retries are a better approach than adding a delay. This way it will return as soon as the variable is available.

That said, upping the attemps to 30 mainly for the first variable to it has up to 30 seconds. Logs should also let us see how fast it takes.