Update translate.ps1 to better handle being unable to obtain metadata setting values (#1405)
* Update translate.ps1

Addressing issue where script fails due to metadata value being set to $null.

Removed $script:outs_dir as it was unused.
Removed $script:is_x86 and replaced with detection of 64bit OS.
Moved obtaining metadata values to after network configuration and before they're needed.
Verifying no metadata variable is null and if one is then logging and restarting to try again.

* Update windows_8_x86_byol.wf.json

Updating win8.1 x86 image

* Revert "Update windows_8_x86_byol.wf.json"

This reverts commit b778886224012af69ca8ba15d975eba298dfb943.

* Update translate.ps1

Changing back to write-host, it should write the message to the log and not return the message.

* Update daisy_workflows/image_import/windows/translate.ps1

Fixing wording of log message.

Co-authored-by: zoran15 <zoranl@google.com>

* Update daisy_workflows/image_import/windows/translate.ps1

Fixing wording of log message.

Co-authored-by: zoran15 <zoranl@google.com>

* Update translate.ps1

Adding retries and a sleeping between each attempt when unable to obtain a value from metadata server.

* Update translate_windows_wf.json

Removing is_x86 variable from metadata as translate.ps1 will no longer use it.

* Update translate.ps1

Restricting to a single reboot if the metadata variables are not available. If the variables are still unavailable after the metadata reboot the import will fail with the specific error.

In further investigating another common is the imports had VMWare tools removed and they were rebooted. So like the other scripts added 1 minute delay as the delay may likely solve the problem. Also since this introduces a minute delay, reducing the sleep between attemps to 1 second.

* Update translate.ps1

Chatted with Tony and he made a great point that the retries are a better approach than adding a delay. This way it will return as soon as the variable is available.

That said, upping the attemps to 30 mainly for the first variable to it has up to 30 seconds. Logs should also let us see how fast it takes.

* Update translate.ps1

Chatted with Eric and he made a great point that the retries are a better approach than adding a delay. This way it will return as soon as the variable is available.

That said, upping the attemps to 30 mainly for the first variable to it has up to 30 seconds. Logs should also let us see how fast it takes.

Co-authored-by: zoran15 <zoranl@google.com>