Replacing 1 minute wait with network connectivity checks and reboot remediation in run_startup_script.cmd.
Removed 1 minute wait.
Added checks for connectivity to metadata service and packages.cloud.google.com with 30 second wait for any failure.
If there are no failures the checks should completed in ~5 seconds and with failures it will wait up to 310 seconds.
There is a final check of metadata service to determine if a reboot should be attempted to remediate the issue.

Also logging to the x64 script that was already in the x86 script.