Add SerialPortsToLog parameter to InstanceBase
When using the WaitForInstancesSignal with a port other than 1:

  "WaitForInstancesSignal": [{
    "Name": "builder",
    "SerialOutput": {
      "Port": 3,
      "SuccessMatch": "BuilderSuccess:",
      "FailureMatch": "BuilderFailure:",
      "StatusMatch": "BuilderStatus:"
    }
  }]

It is not possible to see the entire serial port logs in GCS due to the
fact that the logging is actually an attribute of the CreateInstances
and the port number there is hardcoded to 1.

Introduce SerialPortsToLog parameter for instance base to configure
which serial ports are being logged for the given instance.