Image import: Pick translator based on OS detection when `--os` unspecified  (#1451)
This change modifies the user-facing surface of `gce_vm_image_import` in two ways:

## 1. `--os` is optional

Previously users always had to specify either `--os`, `--data-disk`, or a custom translation workflow. Following this, if a customer omits all three options, then the results of OS detection are used to determine which translation workflow to apply to the imported disk.

## 2. A new flag `--byol`

When running in the new OS-detection mode, if a particular OS is available as either on-demand or BYOL, the on-demand mode is chosen by default. To import as BYOL, the user specifies the `-byol` flag.