Modify scratch bucket creation name to not use "."
Periods can be used in the creation of a GCS bucket, but if it exists
then there is a requirement that the domain that is used with the period
be owned by the project. We can instead modify the project name to use
elgoog_com so that we do not rely on a project owning that domain.

https://cloud.google.com/storage/docs/naming-buckets#example_bucket_names