Update export_disk_ext.sh
Added a more robust test condition and error message for both Storage and Compute endpoints.

This test removes the original curl to the Discovery API, which can be accessed without authentication but tests access to an unrelated API.
Instead, this test now checks HTTP Statuscodes for these endpoints. Without specific parameters, the endpoints will return 4XX; that is acceptable because we're yet not testing permission to reach a specific resource. Returning any status code equal to or greater than 200 indicates that at least the VM can communicate to this endpoint.

I've also changed the error message to be more useful. The previous message over-emphasized Private Google Access, which made troubleshooting difficult.