Ubuntu import: When possible, don't install cloud-init 21.3-1 (#1763)
This updates Ubuntu translate to avoid installing cloud-init 21.3-1, which blocks systemd from starting the guest agent on first boot.

It achieves this by querying apt to find the available package versions, and picking a version such that 21.3-1 is avoided if possible. If a version other than 21.3-1 is found, this falls back to the previous behavior of allowing apt to choose which version to install. In this case, the guest agent will start after the first boot, which is the current behavior.