Daisy: Eagerly record serial logs. (#1779)
While debugging failed imports, I've noticed that a subset of logs don't include serial logs. This occurred since there's a race condition between reading serial port logs, and daisy declaring that the workflow is finished. The go routine that collects serial port logs continues after daisy makes that declaration, to ensure that all of the instances logs get sent to Cloud Logging in a single batch.

This PR ensures that serial port logs are eagerly collected, so that when a workflow finishes, ReadSerialPortLogs returns all of the instance's serial port logs. It achieves this by accumulating each portion of serial port logs as they're available, and enforces an explicit call to write to cloud logging. This allows a client of daisy.Workflow to call ReadSerialPortLogs at any point and retrieve the full view of serial port logs.