CLI Tools: Allow workflows to be re-run (#1777)
This PR updates DaisyWorker to allow workflows to be retried. To achieve this, clients of DaisyWorker supply a closure that generates a workflow. On the first invocation, and any retries, this factory is used to create a new workflow. This is required since daisy workflows aren't designed to be run multiple times.