OVF Import: Delete images when instance creation fails (#1840)
This PR updates OVF import so that temporary images are deleted when instance creation fails. Prior to this, if the final instance couldn't be created, the temporary images would remain in the user's project. There are various reasons the final instance cannot be created; two examples are: insufficient quota, and an invalid machine type flag.