Image Import: Send InspectionResults in OutputInfo (#1833)
OutputInfo already contains the definition for InspectionResults. As the
protos are adjusted, we can safely populate this field in logs.

The field InspectionResults is left in ImageImportParams for backward
compatibility and will be removed in the future.

A manually defined struct InspectionResults in log_entry.go is not used
anywhere and can be safely removed.