Update windows client publish workflows (#1964)
* Update windows client publish workflows

This reworks the logic in the publish workflow to add the testing environment, allowing the publishing to bct-prod, which was missing in the original push. This is needed for integration tests.

* Update image family for Windows Client images

Matches previous iterations' naming schema

* Update image family in Windows 11 publish file

Make image family naming schema consistent with other versions. Windows 11 only has a 64 bit version, but tagging it this way allows for us to distinguish between ARM version if needed in the future.