Kernel Commit Queries
This workflow intends to inspect a Linux system to determine if its
kernel contains required patch(es) implementing a given feature.

The initial version implements a "apply & check" strategy where it
will reconstruct the changes chain between two kernel versions and
check if the expected patches are present in the chain.

Additionally it implements a "changelog" strategy where the kernel
package's changelog is used as checking input.