Add logic to skip Server 2012R2 in W10/11 KB check (#2019)
* Add logic to skip Server 2012R2 in W10/11 KB check

Server 2012R2 doesn't have CurrentMajorVersionNumber or CurrentMinorVersionNumber so this breaks the build for 2012R2. The known issue KB doesn't target this version of Windows Server, so it is safe to skip in this manner.

* Switch to using Environment variables instead of reg keys

This method works across all versions of Server, so no special logic is required if we do it this way.