Remove byol suffix from Windows client workflows (#2085)
Turns out the previous change to add the suffix is incompatible with how the windows build pipeline is generating file paths. I figure it's safe to drop byol as we don't publish any payg client images. Reworking the pipeline to accommodate the suffix would be more trouble than it's worth.