el: fix rsync errors with -boot iso variant (#2117)
The distinction between -boot and -dvd1 isos for what concerns /EFI
and /images directory is the content ownership, in the -boot some files
will be 'nobody' owned, whereas in the dvd1 image all files are 'root'
owned, rsync will try to sync the content attributes but fail for 'nobody'
since the group and user are not yet available.

This change makes it consistent both -boot and -dvd1 use cases where the
resulting/installed content will always be owned by root.