Update build files for Debian 13 (#2452)
* First attempt at updating fai build

I expect this to break, but gotta start somewhere.

* Revert to old cloud_images version; hardcode bookworm

Sid/Trixie haven't been updated in our repos yet so the agent version there is way too old. Doing this for now to workaround it with the bookworm version which should work.

* Re-disable tests

These were disabled for 12, didn't seem to make a difference for 13

* Revert back to suite for apt sources

The hardcode is only needed to test locally util we have the virtual repos set up for Trixie

* Remove DS_Store files

Accidentally committed these when I moved branches

* Delete additional DS_Store

* Update Debian 13 daisy workflow

Replacing Sid with Trixie for stable release name