Update installpackage file check
This will unlink a broken symlink if it cannot be followed, while retaining the behavior to back up the other resolv.conf file if it exists.

Context: Debian-13 uses a symlink for this file which is is broken when the disk is mounted as a secondary disk. This allows us to remove the broken symlink and continue, which should be fine as Debian-13 will rebuild the link/file on reboot.