Update installpackage resolv.conf check (#2480)
* Update installpackage file check

This will unlink a broken symlink if it cannot be followed, while retaining the behavior to back up the other resolv.conf file if it exists.

Context: Debian-13 uses a symlink for this file which is is broken when the disk is mounted as a secondary disk. This allows us to remove the broken symlink and continue, which should be fine as Debian-13 will rebuild the link/file on reboot.

* Expand logic tree for resolv.conf check

I had originally tested with this one but tried a shorter version that failed. Accidentally committed that one before saving the file again.

This is the correct logic to preserve the behavior for previous OSes while accommodating the change in Debian 13.

* Remove trailing comma from comment

This caused a validation error

* Update comment line lengths

Should conform now

* Update filepath for resolv.conf check

Absolute vs relative pathing leaves less room for error