Fix patch didn't apply issue
COS have a TPC patch that's added to the master branch. But the script
didn't support google-guest-agent repo so applying the patch will
complain. Let's add the short-term fix for now to add the patch to the
ignored list.

In the long term, we should migrate the script to add google-guest-agent
repo and always use master branch.