Add new keyring to Trixie config for testing (#2556)
* Fix build error for Trixie

We need this keyring for now it seems

* Remote whitespace

* Add new keyring to Trixie config for testing

This is just to allow the build to run with the new method to verify everything works as intended. The real solution will be to package the keyring so we can update it if the key ever rotates, similar to how we did it for packages.cloud.google.com prior

* Update description for TRIXIE hook

This will probably also end up merged into GCE_SPECIFIC once we transition older builds.

* Update Trixie build with correct URLs

We can move closer to standard now that the redirection is working.